[FEAT] Time is a double value now instead of long. Time unit is changed to second instead of millisecond. - Added javadoc comments to AbstractLogger.java - Add new scenario for simulating Bitcoin Network - Fixed Bitcoin Miner Node

+++ src/main/java/jabs/data/bitcoin/BitcoinBlockWithTx.java
@@ -10,7 +10,7 @@ import static jabs.network.BlockFactory.BITCOIN_BLOCK_HEADER_SIZE;
 public class BitcoinBlockWithTx extends BitcoinBlock {
     private final Set<BitcoinTx> Txs;
 
-    public BitcoinBlockWithTx(int height, long creationTime, BitcoinBlock parent, BitcoinMinerNode creator, Set<BitcoinTx> txs) {
+    public BitcoinBlockWithTx(int height, double creationTime, BitcoinBlock parent, BitcoinMinerNode creator, Set<BitcoinTx> txs) {
         super(0, height, creationTime, creator, parent);
         Txs = txs;
